+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * ENHANCEMENT: Add core support for local tools via `CreateProcess.withToolType`, this helper is part of `Fake.DotNet.Cli` and available after `open Fake.Core` - https://github.com/fsharp/FAKE/pull/2399
 * ENHANCEMENT: Add `ToolType` support in `ReportGenerator` and `Paket`, thanks @SteveGilham and @Tarmil - https://github.com/fsharp/FAKE/pull/2399
 * ENHANCEMENT: Add `FAKE_COREFX_VERBOSE` in order to increase verbosity for the FAKE libraries
+* ENHANCEMENT: Add `DotNet.setupEnv` in order to improve working with installed SDKs (for example when calling fable), fixes https://github.com/fsharp/FAKE/issues/2405
 * BUGFIX: Address https://github.com/fsharp/FAKE/issues/2406 by only printing a warning instead of crashing
 * BUGFIX: Partially address https://github.com/fsharp/FAKE/issues/2401 by allowing the user to continue and request additional info.